Refactoring so the attribute setting logic can be shared
